# apps/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students/{id}/parents/{parentId}": {
      "post": {
        "operationId": "StudentsController_addParentToStudent",
        "parameters": [
          {
            "name": "parentId",
            "required": true,
            "in": "path",
            "description": "ID родителя",
            "schema": {}
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID студента",
            "schema": {}
          }
        ],
        "responses": {
          "201": {
            "description": "Родитель успешно привязан к студенту"
          },
          "404": {
            "description": "Студент или родитель не найден"
          },
          "409": {
            "description": "Родитель уже привязан к этому студенту"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Привязать родителя к студенту",
        "tags": [
          "Students"
        ]
      },
      "delete": {
        "operationId": "StudentsController_removeParentFromStudent",
        "parameters": [
          {
            "name": "parentId",
            "required": true,
            "in": "path",
            "description": "ID родителя",
            "schema": {}
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID студента",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Родитель успешно отвязан от студента"
          },
          "404": {
            "description": "Студент, родитель не найден или связь не существует"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Отвязать родителя от студента",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/parents": {
      "post": {
        "operationId": "ParentsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Родитель успешно зарегистрирован"
          },
          "409": {
            "description": "Пользователь уже является родителем"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Зарегистрировать пользователя как родителя",
        "tags": [
          "Parents"
        ]
      },
      "get": {
        "operationId": "ParentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список всех родителей с информацией о пользователях"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить всех родителей",
        "tags": [
          "Parents"
        ]
      }
    },
    "/parents/user/{userId}": {
      "get": {
        "operationId": "ParentsController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID пользователя",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о родителе"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить запись родителя по ID пользователя",
        "tags": [
          "Parents"
        ]
      }
    },
    "/parents/search": {
      "get": {
        "operationId": "ParentsController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска родителей"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Поиск родителей",
        "tags": [
          "Parents"
        ]
      }
    },
    "/parents/statistics": {
      "get": {
        "operationId": "ParentsController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Статистика родителей"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить статистику родителей",
        "tags": [
          "Parents"
        ]
      }
    },
    "/parents/{id}": {
      "get": {
        "operationId": "ParentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID родителя",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Полная информация о родителе"
          },
          "404": {
            "description": "Родитель не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить родителя по ID",
        "tags": [
          "Parents"
        ]
      },
      "patch": {
        "operationId": "ParentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID родителя",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Данные родителя успешно обновлены"
          },
          "404": {
            "description": "Родитель не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить данные родителя",
        "tags": [
          "Parents"
        ]
      },
      "delete": {
        "operationId": "ParentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID родителя",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Родитель успешно удален"
          },
          "404": {
            "description": "Родитель не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить родителя",
        "tags": [
          "Parents"
        ]
      }
    },
...
    "/payments/parent/{parentId}": {
      "get": {
        "operationId": "PaymentsController_getParentPayments",
        "parameters": [
          {
            "name": "parentId",
            "required": true,
            "in": "path",
            "description": "ID родителя",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Платежи родителя с детальной информацией"
          },
          "404": {
            "description": "Родитель не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все платежи родителя (за всех детей)",
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant ABAI API",
...
    "version": "1.0.0",
    "contact": {
      "name": "ABAI Support Team",
      "url": "https://abai.edu.kz",
      "email": "support@abai.edu.kz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.abai.edu.kz",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен",
        "in": "header"
      }
    },
...
  }
}